+++ package.json
@@ -18,6 +18,7 @@
     "@ckeditor/ckeditor5-list": "*",
     "@ckeditor/ckeditor5-paragraph": "*",
     "@ckeditor/ckeditor5-presets": "*",
+    "@ckeditor/ckeditor5-typing": "*",
     "gulp": "^3.9.0",
     "guppy-pre-commit": "^0.4.0"
   },
